Implement basic lists in C I've implemented a basic framework for C lists. It's certainly not the most optimized thing, but it seems to work - at least for int lists.
One thing to consider is what to do with something invalid happens.
Right now, for example, I have no function for reading from a list,
because there is nothing special to do - I would like to make a
function that reads from a list to prevent the user from doing
invalid reads (such as when the list is empty or if the index is
less than zero) but there is no exception mechanism or anything yet.

Another thing to note is that right now, I don't have a way to
implement, for example, a general 'find' method - I could do one
that worked on primitive types, but it wouldn't work on strings,
because they require a special comparison method.